+++ bot/on_message/on_message.py
@@ -21,7 +21,7 @@ from bot.scripts.delete_message_if_conditions_are_met import (
     reject_artist_text_in_gallery,
     delete_based_images_in_general,
 )
-from bot.scripts.is_newbie import is_newbie
+# from bot.scripts.is_newbie import is_newbie
 from bot.scripts.assert_old_users_have_connected import assert_old_users_have_connected
 from bot.scripts.message.forbidden import (
     name_contains_profanity,
@@ -49,10 +49,48 @@ from datetime import datetime, timezone, timedelta
 import pytz
 from bot.on_message.bots.gptbot import post_gpt_response
 tz = pytz.timezone('America/Los_Angeles')
+import discord
+
+# define a custom class called Message that takes in a discord.Message and adds some attributes to it
+class Message(discord.Message):
+    
+    #initialize the class 
+    def __init__(self, message):
+        self.id = message.id
+        self.type = message.type
+        self.flags = message.flags
+        self.message = message
+        self.content = message.content
+        self.author = message.author
+        self.channel = message.channel
+        self.guild = message.guild
+        self.now = datetime.now(tz)
+        self.raw_mentions = message.raw_mentions
+        self.raw_role_mentions = message.raw_role_mentions
+        self.raw_channel_mentions = message.raw_channel_mentions
+        self.reference = message.reference
+
+        # a variable which holds a random float between 0 and 1
+        self.die_roll = random.random()        
+        self.is_newbie = datetime.now(tz) - self.author.joined_at  < timedelta(days= 7)
+        self.is_question = self.content[-1]=='?'
+        self.mentions_rivers = 'rivers' in self.content.lower()
+        self.firestore_user = None
+        self.id_of_user_being_replied_to = None
+        self.user_score = 0# firestore_user["score"]          
+        self.mentions_cuomputer = None
+        self.test_message = None
+        self.nick = None
+        self.language_code = None
+        self.author_roles = None
+
+    def log(self):
+        print(f"die_roll={ round(self.die_roll, 3)} user_score={self.user_score}, language_code={self.language_code}, is_newbie={self.is_newbie}, is_question={self.is_question}, mentions_rivers={self.mentions_rivers}, mentions_cuomputer={self.mentions_cuomputer}")
+
+
 
 @client.event
 async def on_message(message):
-
     author = message.author
     channel = message.channel
 
@@ -80,7 +118,7 @@ async def on_message(message):
     if await message_is_forbidden(message, role_names):
         return
 
-    if await is_request_for_server_time(message, author):
+    if await is_request_for_server_time(message):
         return    
 
     # Add time-based roles so they can access more channels I MOVED THIS UP
@@ -108,7 +146,8 @@ async def on_message(message):
     await reject_artist_text_in_gallery(message, author_roles)
     # await reject_in_focus_channel(message, author_roles)
 
-    await delete_based_images_in_general(message, author_roles, now)
+
+    # await delete_based_images_in_general(message, author_roles, now)
 
     # print(firestore_user)
     # await add_remove_roles_for_specific_users(author, roles)
@@ -120,35 +159,30 @@ async def on_message(message):
     if await post_library_query_response(nick, message, language_code):
         return
 
+    message = Message(message)
+    message.nick = nick
+    message.language_code = language_code
+    message.test_message = test_message
+    message.author_roles = author_roles
+    message.firestore_user = firestore_user
+    message.user_score = firestore_user["score"]
+    message.id_of_user_being_replied_to = await get_user_id(message)
+    message.mentions_cuomputer  = cuomputer_id in message.raw_mentions or message.id_of_user_being_replied_to == cuomputer_id
 
-    await respond(nick, message, language_code, test_message, author, channel, author_roles, firestore_user)
+    await respond(message, channel)
 
 
-async def respond(nick, message, language_code, test_message, author, channel, author_roles, firestore_user):
+async def respond(message, channel):
     """ Responds to a message with one of various bots """
-
-    # a variable which holds a random float between 0 and 1
-    t = random.random()
-    id_of_user_being_replied_to = await get_user_id(message)
-    is_newbie = datetime.now(tz) - author.joined_at  < timedelta(days= 7)
-    is_question = message.content[-1]=='?'
-    mentions_rivers = 'rivers' in message.content.lower() # i think testmessage doesn't have 'rivers' in it
-    mentions_cuomputer  = cuomputer_id in message.raw_mentions or id_of_user_being_replied_to == cuomputer_id
-    user_score = firestore_user["score"]
-    print(f"t={ round(t, 3)} user_score={user_score}, language_code={language_code}, is_newbie={is_newbie}, is_question={is_question}, mentions_rivers={mentions_rivers}, mentions_cuomputer={mentions_cuomputer}")
-
-    # if meets_conditions_for_standard_response(t, message, newbie):
-    #     print("meets conditions for standard response")
-
+    
+    
+    message.log()
 
     if (
         channel.id in [channels["lounge"]]
-        and user_score > config.gpt_threshold
-        and await post_gpt_response(
-            nick, message, language_code, test_message
-        )
-    ):
-        return
+        and message.user_score > config.gpt_threshold
+        and await post_gpt_response(message)):
+            return
 
     # if t==2 and await post_google_knowledge_response(nick, message, language_code):
     #     return
@@ -163,70 +197,78 @@ async def respond(nick, message, language_code, test_message, author, channel, a
 
     if channel.id == channels["qna"]:
 
-        if await post_roles_response(nick, message, language_code, test_message):
+        if await post_roles_response(message):
             return
 
-        if await post_google_knowledge_response(nick, message, language_code):
+        if await post_google_knowledge_response(message):
             return
 
-        if await post_google_response(nick, message, language_code):
+        if await post_google_response(message):
             return
 
-        await post_qna_default_response(nick, message, language_code)
+        await post_qna_default_response(message)
 
-    elif channel.id in [channels["coach"], channels["vangie"], channels["dan"], channels["geezerville"]]:
-        if channel.id == channels["vangie"]:
-            adjective = "funny and flirtatious"
-        elif channel.id == channels["geezerville"]:
-            adjective = "nostalgic and sweet"
-        else: 
-            adjective = "funny"
-        if user_score > config.gpt_threshold and  (
-            (is_question and mentions_rivers and t > .1) or 
-            ((is_question or mentions_rivers) and t > .95) or
-            mentions_cuomputer or
-             (t >.999)
-        ) and await post_gpt_response(nick, message, language_code, adjective=adjective):
+    elif channel.id == channels["coach"]:        
+        if message.user_score > config.gpt_threshold and  (
+            (message.is_question and message.mentions_rivers and message.die_roll > .1) or 
+            ((message.is_question or message.mentions_rivers) and message.die_roll > .95) or
+            message.mentions_cuomputer or
+             (message.die_roll >.999)
+        ) and await post_gpt_response(message, system="You are Rivers Cuomo from Weezer but you're good at giving advice."):
             return
-
-    elif channel.id == channels["vangie"]:
-        if (
-            (is_question and mentions_rivers and t > .1) or 
-            ((is_question or mentions_rivers) and t > .95) or
-            mentions_cuomputer or
-             (t >.999)
-        ) and await post_gpt_response(nick, message, language_code, adjective="funny"):
+        
+    elif channel.id == channels["dan"]:
+        if message.user_score > config.gpt_threshold and  (
+        (message.is_question and message.mentions_rivers and message.die_roll > .1) or 
+        ((message.is_question or message.mentions_rivers) and message.die_roll > .95) or
+        message.mentions_cuomputer or
+            (message.die_roll >.999)
+        ) and await post_gpt_response(message, system = "you are Rivers Cuomo from Weezer but you have adopted the persona of a hyper-opinionated and knowledgable Weezer fan."):
             return
-
+        
+    elif channel.id == channels["geezerville"]:
+        if message.user_score > config.gpt_threshold and  (
+        (message.is_question and message.mentions_rivers and message.die_roll > .1) or 
+        ((message.is_question or message.mentions_rivers) and message.die_roll > .95) or
+        message.mentions_cuomputer or
+            (message.die_roll >.999)
+        ) and await post_gpt_response(message,system = "you are Rivers Cuomo from Weezer. But you're not trying to act younger than you are (52). You're okay talking about middle-age issues including having kids. You sometimes reference 80's nostalgia. ", adjective="nostalgic and sweet"):
+            return
+        
     elif channel.id == channels["pink"]:
         if (
-            (is_question and mentions_rivers ) or 
-            ((is_question or mentions_rivers) and t > .8) or
-             mentions_cuomputer or
-             (t >.99)
-        ) and await post_gpt_response(nick, message, language_code, adjective="flamboyant"):
+            (message.is_question and message.mentions_rivers ) or 
+            ((message.is_question or message.mentions_rivers) and message.die_roll > .8) or
+             message.mentions_cuomputer or
+             (message.die_roll >.99)
+        ) and await post_gpt_response(message, system="You are a gay version of Rivers Cuomo from Weezer.", adjective="funny"):
             return
-
-    elif message.channel.id in [channels["korean"], channels["japanese"] ]:
+        
+    elif channel.id == channels["vangie"]:
+        if message.user_score > config.gpt_threshold and  (
+        (message.is_question and message.mentions_rivers and message.die_roll > .1) or 
+        ((message.is_question or message.mentions_rivers) and message.die_roll > .95) or
+        message.mentions_cuomputer or
+            (message.die_roll >.999)
+        ) and await post_gpt_response(message, system="You are Rivers Cuomo from Weezer but are very smart and like to flirt with women.", adjective="witty and flirtatious"):
+            return
+        
+    elif channel.id in [channels["korean"], channels["japanese"] ]:
         if (
-        (is_question and mentions_rivers) or 
-        ((is_question) and t > .5) or
-        ((mentions_rivers) and t > .5) or
-         mentions_cuomputer or
-            (t >.6)
-        ) and await post_gpt_response(nick, message, language_code):
+        (message.is_question and message.mentions_rivers) or 
+        ((message.is_question) and message.die_roll > .5) or
+        ((message.mentions_rivers) and message.die_roll > .5) or
+         message.mentions_cuomputer or
+            (message.die_roll >.6)
+        ) and await post_gpt_response(message):
             return
 
-    elif is_newbie and t >.96:
-        await post_gpt_response(nick, message, language_code)
+    elif message.is_newbie and message.die_roll >.96:
+        await post_gpt_response(message)
         return
-
-    # # could make these conditions explicit
-    # elif 
-    #     return
-
-    elif t>.997:
-        await post_riverbot_response(nick, message, language_code)
+    
+    elif message.die_roll>.997:
+        await post_riverbot_response(message)
 
 async def get_user_id(message):
     id_of_user_being_replied_to = None
@@ -250,7 +292,7 @@ async def get_user_id(message):
 #         "To report a bug: please use the relevant channel in the Tech Support section.\n",
 #     ]
 #     # response = finalize_response(reply, language_code, nick)
-#     await message.channel.send("".join(reply))
+#     await channel.send("".join(reply))