+++ bot/scripts/is_request_for_server_time.py
@@ -1,12 +1,12 @@
 from datetime import datetime, timezone
 
 
-async def is_request_for_server_time(message, member):
+async def is_request_for_server_time(message):
     today = datetime.now(timezone.utc)
-    age = today - member.joined_at
+    age = today - message.author.joined_at
     joined = max(age.days, 0)
     if message.content in [".svtime", ".svtime."]:
         await message.channel.send(
-            f"{member.name}, you joined my server {joined} days ago."
+            f"{message.author.name}, you joined my server {joined} days ago."
         )
         return True