+++ bot/setup/init_sessions.py
@@ -2,7 +2,7 @@ from google.cloud import dialogflow
 import sys
 
 sys.path.append("...")  # Adds higher directory to python modules path.
-from config import all_response_channels
+from config import channels, all_response_channels
 from rich import print
 from dotenv import load_dotenv
 import os
@@ -34,22 +34,19 @@ def init_sessions():
     the bot will be posting in"""
 
     sessions = []
-    openai_sessions = {}
+
     for channel in all_response_channels:
+
         this_id = channel
 
+        # for dialogflow
         session = session_client.session_path(os.environ.get("GOOGLE_CLOUD_PROJECT"), this_id)
 
         session_data = {"id": this_id, "session": session}
 
         sessions.append(session_data)
 
-        openai_sessions[channel]= ""
-
-        # openai_sessions[channel]= {}
-
-
-
+    openai_sessions = {id: [] for channel_name, id in channels.items()}
     print(f"openai_sessions={openai_sessions}")
     return sessions, openai_sessions